Matrix Operator is now backend-agnostic
