+++ occamypy/cupy/operator/__init__.py
@@ -1,8 +1,6 @@
-from .matrix import *
 from .signal import *
 
 __all__ = [
-    "Matrix",
     "GaussianFilter",
     "ConvND",
     "ZeroPad",
