+++ occamypy/numpy/__init__.py
@@ -3,7 +3,6 @@ from .operator import *
 
 __all__ = [
     "VectorNumpy",
-    "Matrix",
     "ConvND",
     "GaussianFilter",
     "ZeroPad",
