+++ occamypy/numpy/operator/__init__.py
@@ -1,9 +1,7 @@
-from .matrix import *
 from .signal import *
 from .transform import *
 
 __all__ = [
-    "Matrix",
     "ConvND",
     "GaussianFilter",
     "ZeroPad",
